+++ Assets/Scripts/NEAT/TWEANN/Tests/TestDotProd.cs
@@ -5,23 +5,23 @@ using UnityEngine;
 public class TestDotProd : MonoBehaviour {
 
     TWEANN dotProdTest;
-    double[] inputs;
+    float[] inputs;
 
     public void Start()
     {
         dotProdTest = new TWEANN(2, 1, false, FTYPE.ID, 0);
-        inputs = new double[] { 3, 5 };
+        inputs = new float[] { 3, 5 };
 
 
         Debug.Log("Starting test using inputs");
-        foreach(double d in inputs)
+        foreach(float d in inputs)
         {
             Debug.Log(d);    
         }
 
-        double[] results = dotProdTest.Process(inputs);
+        float[] results = dotProdTest.Process(inputs);
 
-        foreach(double sum in results)
+        foreach(float sum in results)
         {
             Debug.Log("Ending test: result = " + sum);
         }
